+++ Backend/SimulationEngine/SimulationEngine.csproj
@@ -15,6 +15,7 @@
 	
 	<ItemGroup>
 	  <ProjectReference Include="..\..\Shared\Common\Common.csproj" />
+	  <ProjectReference Include="..\ECS\ECS.csproj" />
 	</ItemGroup>
 
 </Project>
