+++ Backend/SimulationEngine/UI/Structures/Enums/ScreenID.cs
@@ -1,4 +1,4 @@
-﻿namespace IdelPog.Frontend.UI.Structures.Enums
+﻿namespace IdelPog.SimulationEngine.UI.Structures.Enums
 {
     public enum ScreenID : byte
     {
