+++ Frontend/Frontend/Content/Service/UITextureResolver.cs
@@ -1,5 +1,5 @@
 ﻿using IdelPog.Common.Repository;
-using IdelPog.Frontend.Rendering.Structures.Enums;
+using IdelPog.Common.Structures;
 using Microsoft.Xna.Framework.Graphics;
 
 namespace IdelPog.Frontend.Content.Service
