+++ Frontend/Frontend/Frontend.csproj
@@ -14,6 +14,6 @@
       <ProjectReference Include="..\..\Shared\Messaging\Messaging.csproj" />
     </ItemGroup>
     <ItemGroup>
-      <Folder Include="UI\Service\" />
+      <Folder Include="Rendering\Structures\Enums\" />
     </ItemGroup>
 </Project>
