+++ Frontend/Frontend/MonoGame/FrontendBootstrapper.cs
@@ -1,4 +1,5 @@
 ﻿using IdelPog.Common.Repository;
+using IdelPog.Common.Structures;
 using IdelPog.Frontend.Content.Service;
 using IdelPog.Frontend.MonoGame.Controllers;
 using IdelPog.Frontend.MonoGame.Converter;
@@ -6,7 +7,6 @@ using IdelPog.Frontend.MonoGame.Listeners;
 using IdelPog.Frontend.MonoGame.Mediator;
 using IdelPog.Frontend.Rendering;
 using IdelPog.Frontend.Rendering.Service;
-using IdelPog.Frontend.Rendering.Structures.Enums;
 using IdelPog.Messaging.Messaging;
 using Microsoft.Xna.Framework.Graphics;
