+++ Frontend/Frontend/MonoGame/RenderingPipeline/Converter/RenderableDTOConverter.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
+using IdelPog.Common.DTO;
 using IdelPog.Frontend.Content.Service;
 using IdelPog.Frontend.Rendering.Structures;
-using IdelPog.Frontend.UI.Structures;
 
 namespace IdelPog.Frontend.MonoGame.Converter
 {
